+++ src/CombatText.cpp
@@ -100,7 +100,7 @@ void CombatText::addMessage(std::string message, FPoint location, int displaytyp
 		c->lifespan = duration;
 		c->displaytype = displaytype;
 
-		c->label->set(c->pos.x, c->pos.y, JUSTIFY_CENTER, VALIGN_BOTTOM, c->text, msg_color[c->displaytype]);
+		c->label->set(static_cast<int>(c->pos.x), static_cast<int>(c->pos.y), JUSTIFY_CENTER, VALIGN_BOTTOM, c->text, msg_color[c->displaytype]);
 		combat_text.push_back(*c);
 		delete c;
 	}
